+++ packages/av-cliper/demo/fast-concat-mp4.ts
@@ -1,4 +1,4 @@
-import { Log } from '../src/log';
+import { Log } from '@webav/internal-utils';
 import { fastConcatMP4, mixinMP4AndAudio } from '../src/mp4-utils';
 import { playOutputStream } from './play-video';