+++ packages/av-cliper/vite.config.ts
@@ -1,6 +1,10 @@
 import { resolve } from 'path';
 import { defineConfig } from 'vite';
+import dts from 'vite-plugin-dts';
+import { externalizeDeps } from 'vite-plugin-externalize-deps';
+
 export default defineConfig({
+  plugins: [dts({ rollupTypes: true }), externalizeDeps()],
   build: {
     sourcemap: true,
     lib: {
@@ -8,4 +12,20 @@ export default defineConfig({
       name: 'av-cliper',
     },
   },
+  test: {
+    browser: {
+      enabled: true,
+      name: 'chrome', // browser name is required
+      headless: true,
+    },
+  },
+  server: {
+    host: '0.0.0.0',
+    port: 6066,
+  },
+  ...(['development', 'test'].includes(process.env.NODE_ENV ?? '')
+    ? {
+        publicDir: resolve(__dirname, '../../doc-site/public'),
+      }
+    : {}),
 });