+++ packages/av-recorder/vite.config.ts
@@ -1,21 +1,21 @@
-import { defineConfig } from 'vitest/config';
 import { resolve } from 'path';
+import { defineConfig } from 'vitest/config';
+import dts from 'vite-plugin-dts';
+import { externalizeDeps } from 'vite-plugin-externalize-deps';
 
 export default defineConfig({
-  test: {
-    environment: 'jsdom',
-    coverage: {
-      provider: 'istanbul', // or 'c8'
-    },
-  },
+  plugins: [dts({ rollupTypes: true }), externalizeDeps()],
   build: {
     sourcemap: true,
     lib: {
       entry: resolve(__dirname, 'src/av-recorder.ts'),
-      name: 'av-recoder',
+      name: 'av-recorder',
     },
-    rollupOptions: {
-      external: ['@webav/av-cliper'],
+  },
+  test: {
+    environment: 'jsdom',
+    coverage: {
+      provider: 'istanbul', // or 'c8'
     },
   },
 });