+++ packages/av-canvas/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @webav/av-canvas
 
+## 1.0.0
+
+### Major Changes
+
+- 8212cb5: feat: v1
+
+### Patch Changes
+
+- Updated dependencies [5631c29]
+- Updated dependencies [8212cb5]
+  - @webav/av-cliper@1.0.0
+  - @webav/internal-utils@1.0.0
+
+## 1.0.0-beta.1
+
+### Patch Changes
+
+- Updated dependencies [5631c29]
+  - @webav/av-cliper@1.0.0-beta.1
+  - @webav/internal-utils@1.0.0-beta.1
+
+## 1.0.0-beta.0
+
+### Major Changes
+
+- feat: v1
+
+### Patch Changes
+
+- Updated dependencies
+  - @webav/internal-utils@1.0.0-beta.0
+  - @webav/av-cliper@1.0.0-beta.0
+
 ## 0.16.5
 
 ### Patch Changes